requirements.txt
fastapi
langchain-groq
uvicorn
aiohttp
pinecone
python-dotenv
httpx
mistralai
langchain-mistralai